Correct readme formatting
It is proper style to wrap lines at 72 characters. However, this does
not mean that lines should be skipped as they are with paragraphs. This
commit fixes the formatting issue, making it easier to read. I am 
creating a PR for this because I want to draw attention to the issue.